Changed routes addresses & seprated links related modules
