+++ resources/links/AddLink.py
@@ -11,27 +11,7 @@ from common.Link import Link
 import validators
 
 
-class Links(Resource):
-    @jwt_required
-    @swag_from('../yml/links_get.yml')
-    def get(self, id=-1):
-        """ If client requests for /links will get whole links;
-        else if requests for /links/[ID] will get specified link.
-        NOTE: -1 is used as sentinel value
-        """
-        current_user_username = get_jwt_identity()
-        links_list = DbHandler.get_links(current_user_username, id)
-        if links_list:
-            return make_response(
-                jsonify(links_list),
-                200
-            )
-        else:
-            return make_response(
-                jsonify(msg="Link not found!"),
-                404
-            )
-
+class AddLink(Resource):
     @jwt_required
     @swag_from('../yml/links_post.yml')
     def post(self):
@@ -73,25 +53,4 @@ class Links(Resource):
         else:
             return make_response(
                 jsonify(msg="Failed to add new link"), 500
-            )
-
-    @jwt_required
-    @swag_from('../yml/links_delete.yml')
-    def delete(self, id):
-        current_user_username = get_jwt_identity()
-
-        remove_status = DbHandler.remove_link(current_user_username, id)
-        if remove_status == "OK":
-            return make_response(
-                jsonify(msg="Link removed successfully.", link_id=id),
-                200
-            )
-        elif remove_status == "USER_IS_NOT_OWNER":
-            return make_response(
-                jsonify(msg="You don't have permission to " +
-                        "remove this link"), 403
-            )
-        elif remove_status == "ID_NOT_FOUND":
-            return make_response(
-                jsonify(msg="Link doesn't exists!"), 404
             )
